Move container ABCs to var_view module
Python 2 is no longer supported, can assume Python 3.
